+++ CHANGELOG.md
@@ -12,6 +12,7 @@ with regard to its JavaScript and TypeScript bindings' APIs and the Ethereum ABI
 - Radicle Link ID claims contract
 - Pool events when a direct stream from a sender to a receiver is updated
 - Pool events when a sender is updated
+- Pool events when a receiver collects funds
 
 ### Changed
 - Replaced vesting contract with one adapted from Melonport